bug fix
logic that checked for existing rosterGroup was checking with indexOf() instead of using insensitive match. that caused some numbers to be ignored. for example, if there was a "(No Name) 42" in the rosterGroup already, and the next roster entry is "(No Name) 4", the indexOf function returns an unwanted value and the roster entry "(No Name) 4" gets ignored.