ignore instructions inside parentheses
sometimes the sales rep includes instructions for a given player like (Lowercase C) or (45 on left sleeve). We don't want this included as part of the player name, so anything enclosed in parentheses that is not a number will be removed. Grad years that are placed within parentheses like this: (2024) will be retained