Update export_prod_file.js
included a fix for any output filenames that would contain a symbol that could inhibit the file system from saving the file properly. as an example, a player number that contains a ":" breaks the file path and launches a save dialog. now all standard symbols are replaced with an underscore for the filename.